+++ crates/ratchet-loader/src/lib.rs
@@ -1,6 +1,8 @@
 mod ggml;
 mod k_quants;
 
+pub use ggml::*;
+
 #[derive(Debug, thiserror::Error)]
 pub enum LoadError {
     #[error("Invalid GGML Format: {0:#x}")]
