+++ crates/ratchet-models/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "ratchet-models"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ratchet-loader = { path = "../ratchet-loader" }
+byteorder.workspace = true
+anyhow.workspace = true
+bytemuck.workspace = true
+thiserror.workspace = true
+derive-new.workspace = true
+log.workspace = true
+
