+++ crates/ratchet-models/src/lib.rs
@@ -1,0 +1,1 @@
+mod whisper;
