+++ crates/ratchet-models/src/whisper.rs
@@ -1,0 +1,100 @@
+use std::io::{BufRead, SeekFrom};
+
+use byteorder::{LittleEndian, ReadBytesExt};
+use ratchet_loader::{GGMLCompatible, GGMLFormat};
+
+pub struct Whisper;
+
+pub struct WhisperGGMLHeader {
+    pub format: GGMLFormat,
+    pub hparams: HParams,
+    pub filters: MelFilters,
+    pub n_tokens: i32,
+}
+
+#[derive(Debug)]
+pub struct HParams {
+    pub n_vocab: i32,
+    pub n_audio_ctx: i32,
+    pub n_audio_state: i32,
+    pub n_audio_head: i32,
+    pub n_audio_layer: i32,
+    pub n_text_ctx: i32,
+    pub n_text_state: i32,
+    pub n_text_head: i32,
+    pub n_text_layer: i32,
+    pub n_mels: i32,
+    pub ftype: i32,
+}
+
+impl HParams {
+    pub fn read<R: BufRead>(reader: &mut R) -> Result<Self, std::io::Error> {
+        let n_vocab = reader.read_i32::<LittleEndian>()?;
+        let n_audio_ctx = reader.read_i32::<LittleEndian>()?;
+        let n_audio_state = reader.read_i32::<LittleEndian>()?;
+        let n_audio_head = reader.read_i32::<LittleEndian>()?;
+        let n_audio_layer = reader.read_i32::<LittleEndian>()?;
+        let n_text_ctx = reader.read_i32::<LittleEndian>()?;
+        let n_text_state = reader.read_i32::<LittleEndian>()?;
+        let n_text_head = reader.read_i32::<LittleEndian>()?;
+        let n_text_layer = reader.read_i32::<LittleEndian>()?;
+        let n_mels = reader.read_i32::<LittleEndian>()?;
+        let ftype = reader.read_i32::<LittleEndian>()?;
+        Ok(Self {
+            n_vocab,
+            n_audio_ctx,
+            n_audio_state,
+            n_audio_head,
+            n_audio_layer,
+            n_text_ctx,
+            n_text_state,
+            n_text_head,
+            n_text_layer,
+            n_mels,
+            ftype,
+        })
+    }
+}
+
+#[derive(Debug)]
+pub struct MelFilters {
+    pub n_mel: i32,
+    pub n_fft: i32,
+    pub mels: Vec<f32>,
+}
+
+impl MelFilters {
+    pub fn read<R: BufRead>(reader: &mut R) -> Result<Self, std::io::Error> {
+        let n_mel = reader.read_i32::<LittleEndian>()?;
+        let n_fft = reader.read_i32::<LittleEndian>()?;
+
+        let mels = (0..(n_mel * n_fft))
+            .map(|_| reader.read_f32::<LittleEndian>())
+            .collect::<Result<Vec<f32>, std::io::Error>>()?;
+
+        Ok(Self { n_mel, n_fft, mels })
+    }
+}
+
+impl GGMLCompatible for Whisper {
+    type ModelHeader = WhisperGGMLHeader;
+
+    fn load_header<R: std::io::prelude::BufRead + std::io::prelude::Seek>(
+        reader: &mut R,
+    ) -> Result<Self::ModelHeader, ratchet_loader::LoadError> {
+        let format = GGMLFormat::read(reader)?;
+        let hparams = HParams::read(reader)?;
+        let filters = MelFilters::read(reader)?;
+        let n_tokens = reader.read_i32::<LittleEndian>()?;
+        for _ in 0..n_tokens {
+            let token_len = reader.read_u32::<LittleEndian>()?;
+            reader.seek(SeekFrom::Current(token_len as i64))?;
+        }
+        Ok(Self::ModelHeader {
+            format,
+            hparams,
+            filters,
+            n_tokens,
+        })
+    }
+}
